==> 2-Pandas-estudos-basico/1-Base de Dados.ipynb <==
# Relatório de Análise I

## Importando a Base de Dados

import pandas as pd

# importando
pd.read_csv('dados/aluguel.csv', sep=';')

dados = pd.read_csv('dados/aluguel.csv', sep=';')

dados

type(dados)

dados.info()

dados.head(10)

## Informações Gerais sobre a Base de Dados 

dados.dtypes

tipos = pd.DataFrame(dados.dtypes)
tipos

tipos_de_dados = pd.DataFrame(dados.dtypes, columns= ['Tipos de Dados'])

tipos_de_dados.columns.name = 'Variáveis'

tipos_de_dados

dados.shape

dados.shape[0]

dados.shape[1]

print('A base de dados apresenta {} registros(imóveis) e {} variáveis'.format(dados.shape[0], dados.shape[1]))

data = [['Fulano', 12, 7.0, True],
        ['Sicrano', 15, 3.5, False],
        ['Beltrano', 18, 9.3, True]]

dados_exercicio = pd.DataFrame(data, columns=['Aluno', 'Idade', 'Nota', 'Aprovado'])
dados_exercicio

tipos_dados_exercicios = pd.DataFrame(dados_exercicio.dtypes, columns=['Tipos de Dados'])
tipos_dados_exercicios.columns.name = 'Variáveis'
tipos_dados_exercicios
