==> 2-Pandas-estudos-basico/3-Imoveis Residenciais.ipynb <==
# Relatório de Análise III

## Imóveis Residenciais

import pandas as pd

# importando
dados = pd.read_csv('dados/aluguel.csv', sep=';')

dados.head(10)

list(dados['Tipo'].drop_duplicates())

residencial = ['Quitinete',
 'Casa',
 'Apartamento',
 'Casa de Condomínio',
 'Casa de Vila']

residencial

dados.head(10)

selecao = dados['Tipo'].isin(residencial)
selecao

dados_residencial = dados[selecao]

dados_residencial

list(dados_residencial['Tipo'].drop_duplicates())

## Organizando a Visualização

dados_residencial.shape[0]

dados.shape[0]

dados_residencial.index = range(dados_residencial.shape[0])

dados_residencial

## Exportando a Base de Dados

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep=';')

dados_resiencial_2 = pd.read_csv('dados/aluguel_residencial.csv', sep=';')

dados_resiencial_2

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep=';', index=False)

dados_resiencial_2 = pd.read_csv('dados/aluguel_residencial.csv', sep=';')

dados_resiencial_2

## Exercício 1

import pandas as pd

numeros = [i for i in range(11)]
letras = [chr(i + 65) for i in range(11)]
nome_coluna = ['N']

df = pd.DataFrame(data = numeros, index = letras, columns = nome_coluna)
df

selecao = df['N'].isin([i for i in range(11) if i % 2 == 0])
df = df[selecao]
df
