==> 2-Pandas-estudos-basico/5-Tratamento de Dados Faltantes.ipynb <==
# Relatório de Análise V

## Tratamento de Dados Faltantes

import pandas as pd

dados = pd.read_csv('dados/aluguel_residencial.csv', sep=';')

dados.head(10)

dados.isnull()

dados.notnull()

dados.info()

dados['Valor'].isnull()

dados[dados['Valor'].isnull()]

A = dados.shape[0]
dados.dropna(subset = ['Valor'], inplace=True)
B = dados.shape[0]
A - B

dados[dados['Valor'].isnull()]

## Tratamento de Dados Faltantes (continuação)

dados[dados['Condominio'].isnull()].shape[0]

selecao = (dados['Tipo'] == 'Apartamento') & (dados['Condominio'].isnull())

A = dados.shape[0]
dados = dados[~selecao]
B = dados.shape[0]
A - B

dados[dados['Condominio'].isnull()].shape[0]

dados = dados.fillna({'Condominio': 0, 'IPTU': 0})

dados[dados['Condominio'].isnull()].shape[0]

dados[dados['IPTU'].isnull()].shape[0]

dados.info()

dados.to_csv('dados/aluguel_residencial.csv', sep=';', index=False)

## Exercício I

# Utilizando os recursos aprendidos nesta aula, julgue o conjunto de operações de limpeza do DataFrame abaixo:
imoveis = pd.DataFrame([['Apartamento', None, 970, 68], 
                        ['Apartamento', 2000, 878, 112], 
                        ['Casa', 5000, None, 500], 
                        ['Apartamento', None, 1010, 170], 
                        ['Apartamento', 1500, 850, None], 
                        ['Casa', None, None, None], 
                        ['Apartamento', 2000, 878, None], 
                        ['Apartamento', 1550, None, 228], 
                        ['Apartamento', 2500, 880, 195]], 
                        columns = ['Tipo', 'Valor', 'Condominio', 'IPTU'])

imoveis

# 1) Elimina os registros que não apresentam a variável Valor:
imoveis.dropna(subset = ['Valor'], inplace = True)
imoveis

# 2) Elimina os imóveis do tipo Apartamento que não apresentam valor Condominio:
selecao = (imoveis['Tipo'] == 'Apartamento') & (imoveis['Condominio'].isnull())
imoveis = imoveis[~selecao]
imoveis

# 3) Substitui os valores faltantes que restam nas variáveis Condominio e IPTU por zero:
imoveis = imoveis.fillna({'Condominio': 0, 'IPTU': 0})
imoveis

# 4) Reconstrói o índice do DataFrame resultante:
imoveis.index = range(imoveis.shape[0])
imoveis
