==> 2-Pandas-estudos-basico/6-Criando Novas Variaveis.ipynb <==
# Relatório de Análise VI

## Criando Novas Variáveis

import pandas as pd

dados =pd.read_csv('dados/aluguel_residencial.csv', sep=';')

dados.head(10)

dados.head()

dados['Valor Bruto'] = dados['Valor'] + dados['Condominio'] + dados['IPTU']

dados.head(10)

dados['Valor m2'] = dados['Valor'] / dados['Area']

dados.head(10)

dados['Valor m2'] = dados['Valor m2'].round(2)

dados.head(10)

dados['Valor Bruto m2'] = (dados['Valor Bruto'] / dados['Area']).round(2)

dados.head(10)

casa = ['Casa', 'Casa de Condomínio', 'Casa de Vila']

dados['Tipo Agregado'] = dados['Tipo'].apply(lambda x: 'Casa' if x in casa else 'Apartamento')

dados

## Excluir Variáveis 

dados_aux = pd.DataFrame(dados[['Tipo Agregado', 'Valor m2', 'Valor Bruto', 'Valor Bruto m2']])

dados_aux.head(10)

del dados_aux['Valor Bruto']

dados_aux.head(10)

dados_aux.pop('Valor Bruto m2')

dados_aux

dados_coluna = dados.drop(['Valor Bruto', 'Valor Bruto m2'], axis=1)

dados_coluna

dados_linha = dados.drop([0, 1, 2, 3])

dados_linha

 dados.drop(['Valor Bruto', 'Valor Bruto m2'], axis=1, inplace=True)

dados.head(10)

dados.to_csv('dados/aluguel_residencial.csv', sep=';', index = False)

## Exercício I

import pandas as pd
alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas'])
alunos
